+++ src/components/Path/OAPathEndpoint.vue
